backend/requirements.txt
# Medallion Architecture Backend Dependencies

# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
supabase==2.3.0
postgrest==0.13.2

# HTTP Clients
httpx>=0.24.0,<0.25.0  # Compatible with supabase 2.3.0
requests==2.31.0

# Environment
python-dotenv==1.0.0

# Utilities
python-multipart==0.0.6

# Logging
python-json-logger==2.0.7

# Optional: For development
# pytest==7.4.3
# pytest-asyncio==0.21.1

